IGNITE-13071 Rework cache tests structure.
